Copy tags from the AWS service instance to Consul
Since tags in AWS are a pair "key": "value", both parts are concatenated
before being saved in the Consul catalog. When a tag gets updated on AWS
the associated service on Consul gets updated in place with the new
value to false triggers.

This only sync tags from ECS tasks but it should be easy to expand the
scope as needed.

The synchronisation only goes from AWS to Consul as I'm not sure there
is a place to store additional tags to a service instance in Cloud Map.